Fix some more problems with messages view.

+++ admin/src/View/CWMMessages/HTMLView.php
@@ -12,13 +12,15 @@ namespace CWM\Component\Proclaim\Administrator\View\CWMMessages;
 
 defined('_JEXEC') or die;
 
+use CWM\Component\Proclaim\Administrator\Extension\ProclaimComponent;
 use CWM\Component\Proclaim\Administrator\Helper\CWMProclaimHelper;
+use Joomla\CMS\Component\ComponentHelper;
 use Joomla\CMS\Factory;
-use Joomla\CMS\HTML\HTMLHelper;
 use Joomla\CMS\Language\Multilanguage;
 use Joomla\CMS\Language\Text;
 use Joomla\CMS\MVC\View\GenericDataException;
 use Joomla\CMS\MVC\View\HtmlView as BaseHtmlView;
+use Joomla\CMS\Plugin\PluginHelper;
 use Joomla\CMS\Toolbar\Toolbar;
 use Joomla\CMS\Toolbar\ToolbarHelper;
 use Joomla\Component\Content\Administrator\Extension\ContentComponent;
@@ -33,70 +35,6 @@ use Joomla\Component\Content\Administrator\Helper\ContentHelper;
 class HTMLView extends BaseHtmlView
 {
 	/**
-	 * Can Do
-	 *
-	 * @var object
-	 * @since    7.0.0
-	 */
-	public object $canDo;
-
-	/**
-	 * Books
-	 *
-	 * @var object
-	 * @since    7.0.0
-	 */
-	public object $books;
-
-	/**
-	 * Teachers
-	 *
-	 * @var object
-	 * @since    7.0.0
-	 */
-	public object $teachers;
-
-	/**
-	 * Series
-	 *
-	 * @var object
-	 * @since    7.0.0
-	 */
-	public object $series;
-
-	/**
-	 * Message Types
-	 *
-	 * @var object
-	 * @since    7.0.0
-	 */
-	public object $messageTypes;
-
-	/**
-	 * Years
-	 *
-	 * @var object
-	 * @since    7.0.0
-	 */
-	public object $years;
-
-	/**
-	 * Filter Levels
-	 *
-	 * @var array
-	 * @since    7.0.0
-	 */
-	public array $f_levels;
-
-	/**
-	 * Side Bar
-	 *
-	 * @var object
-	 * @since    7.0.0
-	 */
-	public object $sidebar;
-
-	/**
 	 * Items
 	 *
 	 * @var array
@@ -107,18 +45,18 @@ class HTMLView extends BaseHtmlView
 	/**
 	 * Pagination
 	 *
-	 * @var object
+	 * @var      \Joomla\CMS\Pagination\Pagination
 	 * @since    7.0.0
 	 */
-	protected object $pagination;
+	protected $pagination;
 
 	/**
 	 * State
 	 *
-	 * @var object
+	 * @var  \Joomla\CMS\Object\CMSObject
 	 * @since    7.0.0
 	 */
-	protected object $state;
+	protected $state;
 
 	/**
 	 * Active Filters
@@ -131,10 +69,26 @@ class HTMLView extends BaseHtmlView
 	/**
 	 * Filters of the Form
 	 *
-	 * @var mixed ?still not setup yet?
+	 * @var  \Joomla\CMS\Form\Form
 	 * @since    7.0.0
 	 */
 	public $filterForm;
+
+	/**
+	 * All transition, which can be executed of one if the items
+	 *
+	 * @var  array
+	 * @since 4.0.0
+	 */
+	protected array $transitions = [];
+
+	/**
+	 * Is this view an Empty State
+	 *
+	 * @var   boolean
+	 * @since 4.0.0
+	 */
+	private $isEmptyState = false;
 
 	/**
 	 * Execute and display a template script.
@@ -143,42 +97,31 @@ class HTMLView extends BaseHtmlView
 	 *
 	 * @return  void  A string if successful, otherwise a JError object.
 	 *
-	 * @see     fetch()
-	 * @since   11.1
 	 * @throws  \Exception
+	 * @since   11.1
+	 * @see     fetch()
 	 */
 	public function display($tpl = null)
 	{
-		$this->items      = $this->get('Items');
-		$this->pagination = $this->get('Pagination');
-		$this->state      = $this->get('State');
-
-		$this->canDo = CWMProclaimHelper::getActions('', 'message');
-		$modelView   = $this->getModel();
-
+		$this->items         = $this->get('Items');
+		$this->pagination    = $this->get('Pagination');
+		$this->state         = $this->get('State');
+		$this->canDo         = CWMProclaimHelper::getActions('', 'message');
 		$this->filterForm    = $this->get('FilterForm');
 		$this->activeFilters = $this->get('ActiveFilters');
 
-		// Check for errors
-		if (count($errors = $this->get('Errors')))
+		if (ComponentHelper::getParams('com_proclaim')->get('workflow_enabled'))
+		{
+			PluginHelper::importPlugin('workflow');
+
+			$this->transitions = $this->get('Transitions');
+		}
+
+		// Check for errors.
+		if (\count($errors = $this->get('Errors')) || $this->transitions === false)
 		{
 			throw new GenericDataException(implode("\n", $errors), 500);
 		}
-
-		// Levels filter.
-		$options   = array();
-		$options[] = HTMLHelper::_('select.option', '1', Text::_('J1'));
-		$options[] = HTMLHelper::_('select.option', '2', Text::_('J2'));
-		$options[] = HTMLHelper::_('select.option', '3', Text::_('J3'));
-		$options[] = HTMLHelper::_('select.option', '4', Text::_('J4'));
-		$options[] = HTMLHelper::_('select.option', '5', Text::_('J5'));
-		$options[] = HTMLHelper::_('select.option', '6', Text::_('J6'));
-		$options[] = HTMLHelper::_('select.option', '7', Text::_('J7'));
-		$options[] = HTMLHelper::_('select.option', '8', Text::_('J8'));
-		$options[] = HTMLHelper::_('select.option', '9', Text::_('J9'));
-		$options[] = HTMLHelper::_('select.option', '10', Text::_('J10'));
-
-		$this->f_levels = $options;
 
 		// We don't need toolbar in the modal window.
 		if ($this->getLayout() !== 'modal')
@@ -192,9 +135,23 @@ class HTMLView extends BaseHtmlView
 				$this->filterForm->removeField('language', 'filter');
 			}
 		}
+		else
+		{
+			// In article associations modal we need to remove language filter if forcing a language.
+			// We also need to change the category filter to show show categories with All or the forced language.
+			if ($forcedLanguage = Factory::getApplication()->input->get('forcedLanguage', '', 'CMD'))
+			{
+				// If the language is forced we can't allow to select the language, so transform the language selector filter into a hidden field.
+				$languageXml = new \SimpleXMLElement('<field name="language" type="hidden" default="' . $forcedLanguage . '" />');
+				$this->filterForm->setField($languageXml, 'filter', true);
 
-		// Set the document
-		$this->setDocument();
+				// Also, unset the active language filter so the search tools is not open by default with this filter.
+				unset($this->activeFilters['language']);
+
+				// One last changes needed is to change the category filter to just show categories with All language or with the forced language.
+				$this->filterForm->setFieldAttribute('category_id', 'language', '*,' . $forcedLanguage, 'filter');
+			}
+		}
 
 		// Display the template
 		parent::display($tpl);
@@ -211,7 +168,7 @@ class HTMLView extends BaseHtmlView
 	protected function addToolbar()
 	{
 		$canDo = ContentHelper::getActions('com_proclaim');
-		$user = Factory::getApplication()->getIdentity();
+		$user  = $this->getCurrentUser();
 
 		// Get the toolbar object instance
 		$toolbar = Toolbar::getInstance('toolbar');
@@ -223,29 +180,70 @@ class HTMLView extends BaseHtmlView
 			$toolbar->addNew('cwmmessage.add');
 		}
 
-		$dropdown = $toolbar->dropdownButton('status-group')
-			->text('JTOOLBAR_CHANGE_STATUS')
-			->toggleSplit(false)
-			->icon('icon-ellipsis-h')
-			->buttonClass('btn btn-action')
-			->listCheck(true);
-		$childBar = $dropdown->getChildToolbar();
-
-		if ($canDo->get('core.edit'))
+		if (!$this->isEmptyState && ($canDo->get('core.edit.state') || \count($this->transitions)))
 		{
-			$toolbar->edit('cwmmessage.edit');
+			$dropdown = $toolbar->dropdownButton('status-group')
+				->text('JTOOLBAR_CHANGE_STATUS')
+				->toggleSplit(false)
+				->icon('icon-ellipsis-h')
+				->buttonClass('btn btn-action')
+				->listCheck(true);
+
+			$childBar = $dropdown->getChildToolbar();
+
+			if (\count($this->transitions))
+			{
+				$childBar->separatorButton('transition-headline')
+					->text('COM_PROCLAIM_RUN_TRANSITIONS')
+					->buttonClass('text-center py-2 h3');
+
+				$cmd      = "Joomla.submitbutton('cwmmessages.runTransition');";
+				$messages = "{error: [Joomla.JText._('JLIB_HTML_PLEASE_MAKE_A_SELECTION_FROM_THE_LIST')]}";
+				$alert    = 'Joomla.renderMessages(' . $messages . ')';
+				$cmd      = 'if (document.adminForm.boxchecked.value == 0) { ' . $alert . ' } else { ' . $cmd . ' }';
+
+				foreach ($this->transitions as $transition)
+				{
+					$childBar->standardButton('transition')
+						->text($transition['text'])
+						->buttonClass('transition-' . (int) $transition['value'])
+						->icon('icon-project-diagram')
+						->onclick('document.adminForm.transition_id.value=' . (int) $transition['value'] . ';' . $cmd);
+				}
+
+				$childBar->separatorButton('transition-separator');
+			}
+
+			if ($canDo->get('core.edit.state'))
+			{
+				$childBar->publish('cwmmessages.publish')->listCheck(true);
+
+				$childBar->unpublish('cwmmessages.unpublish')->listCheck(true);
+
+				$childBar->archive('cwmmessages.archive')->listCheck(true);
+
+				$childBar->checkin('cwmmessages.checkin')->listCheck(true);
+
+				if ($this->state->get('filter.published') != ProclaimComponent::CONDITION_TRASHED)
+				{
+					$childBar->trash('cwmmessages.trash')->listCheck(true);
+				}
+			}
+
+			// Add a batch button
+			if ($user->authorise('core.create', 'com_proclaim')
+				&& $user->authorise('core.edit', 'com_proclaim')
+				&& $user->authorise('core.execute.transition', 'com_proclaim')
+			)
+			{
+				$childBar->popupButton('batch')
+					->text('JTOOLBAR_BATCH')
+					->selector('collapseModal')
+					->listCheck(true);
+			}
 		}
 
-		if ($canDo->get('core.edit.state'))
-		{
-			$toolbar->divider();
-			$toolbar->publish('cwmmessages.publish');
-			$toolbar->unpublish('cwmmessages.unpublish');
-			$toolbar->divider();
-			$toolbar->archive('cwmmessages.archive');
-		}
-
-		if ($this->state->get('filter.published') === ContentComponent::CONDITION_TRASHED && $canDo->get('core.delete'))
+		if (!$this->isEmptyState && $this->state->get('filter.published') == ContentComponent::CONDITION_TRASHED && $canDo->get('core.delete'))
 		{
 			$toolbar->delete('cwmmessages.delete')
 				->text('JTOOLBAR_EMPTY_TRASH')
@@ -253,60 +251,11 @@ class HTMLView extends BaseHtmlView
 				->listCheck(true);
 		}
 
-		if ($this->state->get('filter.published') !== ContentComponent::CONDITION_TRASHED)
+		if ($user->authorise('core.admin', 'com_proclaim') || $user->authorise('core.options', 'com_proclaim'))
 		{
-			$childBar->trash('articles.trash')->listCheck(true);
+			$toolbar->preferences('com_proclaim');
 		}
 
-		// Add a batch button
-		if ($user->authorise('core.create', 'com_proclaim')
-			&& $user->authorise('core.edit', 'com_proclaim')
-			&& $user->authorise('core.edit.state', 'com_proclaim'))
-		{
-			$childBar->popupButton('batch')
-				->text('JTOOLBAR_BATCH')
-				->selector('collapseModal')
-				->listCheck(true);
-		}
-
-		$toolbar->help('JHELP_CONTENT_ARTICLE_MANAGER');
-	}
-
-	/**
-	 * Add the page title to browser.
-	 *
-	 * @return void
-	 *
-	 * @since    7.1.0
-	 */
-	protected function setDocument()
-	{
-		$app = Factory::getApplication();
-		$document = $app->getDocument();
-		$document->setTitle(Text::_('JBS_TITLE_STUDIES'));
-	}
-
-	/**
-	 * Returns an array of fields the table can be sorted by
-	 *
-	 * @return  array  Array containing the field name to sort by as the key and display text as value
-	 *
-	 * @since   3.0
-	 */
-	protected function getSortFields()
-	{
-		return array(
-			'study.studydate'          => Text::_('JBS_CMN_STUDYDATE'),
-			'study.studytitle'         => Text::_('JBS_CMN_TITLE'),
-			'teacher.teachername'      => Text::_('JBS_CMN_TEACHERS'),
-			'series.series_text'       => Text::_('JBS_CMN_SERIES'),
-			'study.ordering'           => Text::_('JGRID_HEADING_ORDERING'),
-			'study.published'          => Text::_('JPUBLISHED'),
-			'study.id'                 => Text::_('JGRID_HEADING_ID'),
-			'access_level'             => Text::_('JGRID_HEADING_ACCESS'),
-			'study.language'           => Text::_('JGRID_HEADING_LANGUAGE'),
-			'locations.location_text'  => Text::_('JBS_CMN_LOCATIONS'),
-			'messageType.message_type' => Text::_('JBS_CMN_MESSAGETYPE')
-		);
+		$toolbar->help('Messages', true);
 	}
 }
